fix: ignore package-lock, rename helpers
