+++ .gitignore
@@ -11,6 +11,7 @@ types/
 
 # Dependencies
 node_modules/
+package-lock.json
 
 # Coverage
 coverage
