+++ packages/graphback-cli/src/helpers/generate.ts
